rename app to a4forms
