--- adhocracy4/datetimefield/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DatetimefieldConfig(AppConfig):
-    name = 'adhocracy4.datetimefield'
-    label = 'a4datetimefield'
