+++ adhocracy4/forms/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FormsConfig(AppConfig):
+    name = 'adhocracy4.forms'
+    label = 'a4forms'
