+++ tests/project/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = (
     'adhocracy4.actions.apps.ActionsConfig',
     'adhocracy4.follows.apps.FollowsConfig',
     'adhocracy4.filters.apps.FiltersConfig',
-    'adhocracy4.datetimefield.apps.DatetimefieldConfig',
+    'adhocracy4.forms.apps.FormsConfig',
 
     # adhocrayc4 generic apps
     'adhocracy4.ratings.apps.RatingsConfig',
